Consolidate fx-to-onnx logic into FxOnnxInterpreter class
Previous to this PR, FX-to-ONNX logic was spread across util functions
in different files: `_export_fx_to_onnx`, `export_fx_to_onnxscript`,
`export_fx_node_to_onnxscript` and `OnnxDispatcher.dispatch`.

* This commit consolidates all utils into a single class
`FxOnnxInterpreter`, encapsulating all related code into a cohesive
abstraction.
https://www.jmgundersen.net/blog/utility-and-helper-classes-a-code-smell
exaplins the issues with having utility classes, which are directly
applicable to util functions.

* `OnnxDispatcher` class was replaced by `FxOnnxInterpreter`.
`OnnxDispatcher.dispatch` was replaced by `FxOnnxINterpreter.step`.
While `OnnxDispatcher.dispatch` only handled `call_function` implementation,
through ONNX Script lokkup, `FxOnnxInterpreter.step` will handle all FX Nodes.
(call_function, call_module, call_method, get_attr, placeholder,
output). This PR does not use `torch.fx.Interpreter` nor does make any other change
to the fx-to-onnx path besides the consolidation into a class.

Additionally, because `_export_fx_to_onnx` also had non fx-to-onnx code,
this commit moved unrelated tasks to separate files:

  * passes (composition, functionalization, etc) were moved to a
`pre_export_passes` until the Pass functionality is properly
implemented. At least now they are isolated in their own place.

* Protobuf serialization was moved to `Exporter.export`
