+++ torch/onnx/_internal/fx/op_validation.py
@@ -13,8 +13,7 @@ import torch
 import torch.fx
 from torch.onnx import _constants, _type_utils
 from torch.onnx._internal import _beartype, onnx_proto_utils
-from torch.onnx._internal.fx import diagnostics
-from torch.onnx._internal.fx.passes import fx_to_onnxscript
+from torch.onnx._internal.fx import diagnostics, dispatcher
 from torch.utils import _pytree
 
 
@@ -30,7 +29,7 @@ def validate_op_between_ort_torch(
 
     The function will run the op in ONNX Runtime and PyTorch and compare the
     results. It doesn't break the exporting process, but saves each op validated
-    result into SARIF, under the section of `fx_to_onnxscript` pass.
+    result into SARIF, under the section of `dispatcher`.
 
     There are three signs can be found:
     1. Blue: Pass
@@ -95,7 +94,7 @@ def validate_op_between_ort_torch(
             else x
             for x in torch_args
         ]
-        kwargs_onnx = fx_to_onnxscript.filter_incompatible_and_dtype_convert_kwargs(
+        kwargs_onnx = dispatcher.filter_incompatible_and_dtype_convert_kwargs(
             torch_kwargs
         )
         try:
