+++ torch/onnx/_internal/fx/passes/__init__.py
@@ -1,11 +1,9 @@
 from .decomp import Decompose
 from .functionalization import Functionalize, RemoveInputMutation
-from .fx_to_onnxscript import export_fx_to_onnxscript
 from .shape_inference import ShapeInferenceWithFakeTensor
 from .virtualization import MovePlaceholderToFront, ReplaceGetAttrWithPlaceholder
 
 __all__ = [
-    "export_fx_to_onnxscript",
     "Decompose",
     "Functionalize",
     "MovePlaceholderToFront",
